Moved out field mapping into a separate data mapping file
